CO-2947 the  Sharing team states
